Tests/Util: move test classes into subdirectories
Move test classes into subdirectories mirroring the name of the class holding the code being tested.

Having the tests for each class in their own subdirectory should make it more straight forward to see what code is covered by tests.
