+++ tests/Core/Util/Common/StripColorsTest.php
@@ -7,7 +7,7 @@
  * @license   https://github.com/PHPCSStandards/PHP_CodeSniffer/blob/master/licence.txt BSD Licence
  */
 
-namespace PHP_CodeSniffer\Tests\Core\Util;
+namespace PHP_CodeSniffer\Tests\Core\Util\Common;
 
 use PHP_CodeSniffer\Util\Common;
 use PHPUnit\Framework\TestCase;
